refactor: move post related files to component/feature/post
