+++ apps/web/src/app/(app)/profile/[handle]/_components/ProfilePosts.tsx
@@ -7,7 +7,7 @@ import { useEffect } from 'react';
 import { useGetUserPostsInfinite } from '@/api/__generated__/post/post';
 import { useGetProfileByHandle } from '@/api/__generated__/profile/profile';
 import type { GetPostsResponsePostsNodesItemContent } from '@/api/__generated__/types';
-import { BaseViewer } from '@/components/ui/editor';
+import { BaseViewer } from '@/components/feature/post';
 import { Skeleton } from '@/components/ui/skeleton';
 import { Spinner } from '@/components/ui/spinner';
