+++ apps/web/src/components/feature/post/PostCard.tsx
@@ -1,0 +1,1 @@
+// TODO: Post viewer
