+++ apps/web/src/components/feature/post/editor/components/EditorTemplates.tsx
@@ -1,3 +1,6 @@
+'use client';
+
+import { useTranslations } from 'next-intl';
 import DEEP_DIVE from 'public/assets/templates/deep-dive.json';
 import PROBLEM_AND_SOLUTION from 'public/assets/templates/problem-and-solution.json';
 import TIL from 'public/assets/templates/til.json';
@@ -30,7 +33,7 @@ const ALL_TEMPLATES: Template[] = [
   WEEKLY_REFLECTION,
 ];
 
-export function getTemplates(locale: string): LocalizedTemplate[] {
+function getTemplates(locale: string): LocalizedTemplate[] {
   return ALL_TEMPLATES.filter(
     (template) =>
       template.label[locale] &&
@@ -44,4 +47,30 @@ export function getTemplates(locale: string): LocalizedTemplate[] {
   }));
 }
 
-export type { LocalizedTemplate };
+interface EditorTemplatesProps {
+  locale: string;
+  onSelect: (template: LocalizedTemplate) => void;
+}
+
+export function EditorTemplates({ locale, onSelect }: EditorTemplatesProps) {
+  const t = useTranslations('components.editor');
+
+  return (
+    <div className="mt-4 text-muted-foreground">
+      {t('templates.suggestions')}
+
+      <div className="flex flex-col gap-2 items-start mt-2">
+        {getTemplates(locale).map((template) => (
+          <button
+            key={template.id}
+            type="button"
+            className="px-3 py-1.5 text-sm text-muted-foreground hover:underline"
+            onClick={() => onSelect(template)}
+          >
+            {template.label}
+          </button>
+        ))}
+      </div>
+    </div>
+  );
+}
