+++ apps/web/src/components/feature/post/editor/primitives/separator/separator.scss
