+++ apps/web/src/components/feature/post/editor/primitives/toolbar/toolbar.scss
