+++ apps/web/src/components/feature/post/index.tsx
@@ -1,0 +1,2 @@
+// TODO: Need to remove this file
+export * from './base';
