--- apps/web/src/components/ui/editor/index.tsx
@@ -1,1 +1,0 @@
-export * from './base';
